tidy api 2.0

+++ src/cn/jpush/api/BaseClient.java
@@ -3,22 +3,21 @@ package cn.jpush.api;
 import java.util.HashSet;
 import java.util.Set;
 
-public  class  BaseClient {
-	protected String masterSecret  = "";
+public class BaseClient {
+	protected String masterSecret = "";
 	protected String appKey = "";
-	protected String sendDescription = "";//发送的描述
-	protected int timeToLive = -1;     //保存离线的时长。秒为单位。默认为保存1天的离线消息
+	protected String sendDescription = "";                              //发送的描述
+	protected long timeToLive = -1;                                      //保存离线的时长。秒为单位。默认为保存1天的离线消息
 	protected boolean enableSSL = false;
-	protected Set<DeviceEnum> devices = new HashSet<DeviceEnum>();//默认发送android和ios
+	protected Set<DeviceEnum> devices = new HashSet<DeviceEnum>();      //默认发送android和ios
 
 	public String getMasterSecret() {
 		return masterSecret;
 	}
 
-	public int getTimeToLive() {
+	public long getTimeToLive() {
 		return timeToLive;
 	}
-
 
 	protected String getAppKey() {
 		return this.appKey;
@@ -49,6 +48,5 @@ public  class  BaseClient {
 	public void setEnableSSL(boolean enableSSL) {
 		this.enableSSL = enableSSL;
 	}
-
 
 }
